citation/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
python-dotenv>=1.0.0
google-generativeai>=0.3.1
streamlit>=1.28.2
typing-extensions>=4.8.0
starlette>=0.27.0
python-multipart>=0.0.6